ID-297 Fixed Issue with expected type of time_of_day

+++ src/schemas.py
@@ -13,7 +13,7 @@ class SimulationRequest(BaseModel):
     temperature: float
     category: str
     day: str
-    time_of_day: int
+    time_of_day: float
 
 class OptimisationRequest(BaseModel):
     """
